Shift `zod` helper types to `@prairielearn/zod`
